tls: Remove the `__serial` field
This has been designed to check for reused TLS keys, but it turns out to be
unnecessary, because each TLS value that has been set also owns a reference
count of the associated key, which prevents the key from being deallocated,
even when the key itself is marked deleted. Therefore TLS keys can't be reused
unless there is no value associated with it.