c11: Implement inline aliases in a more 'proper' way
For Clang, the `__asm__` attribute can alias function declarations,
which affects inline definitions as well; an inline wrapper shall
not be provided, otherwise infinite recursion may occur
(https://github.com/llvm/llvm-project/issues/58724).

For GCC and ICC, this has effects only when a call to the external
definition is made; an inline wrapper should be provided in case it
is needed.

This closes GitHub #73.