sem: Fix comments
The value of a semaphore is decreased when a thread starts waiting, and
increased when a thread is notified. The wait counter of another type of
object (mutex, once flag, condition variable, event) is increased when a
thread starts waiting, and decreased when a thread is notified.

(cherry picked from commit 5b5ddc919bb4994a95eb0ca30febff765112317d)
Signed-off-by: LIU Hao <lh_mouse@126.com>