xglobals: Fix assertion about size of shared memory
`NtCreateSection()` may open an existent section, created by a previous version
of this library. In that case, the region may be smaller than the new struct.
In this case, it is not acceptable to fail. The assertion only makes sense when
the section has just been created.

(cherry picked from commit 18df92e5a9bbfdbeba6f7297a9bba302893d80ce)
Signed-off-by: LIU Hao <lh_mouse@126.com>