gthr_aux: Rename `do_call_once_seh_take_over` to `__MCF_do_call_once_seh_take_over`
Although this function is not declared `.globl`, it seems that `.scl 2` has
the same effect. The reason why we are not willing to declare it `.scl 3` is
that GCC emits `.scl 2` if the function is called, and we don't want to
introduce an inconsistency.

So the function shall have a reserved name in case it escapes the current
object file from a static library.