cxx11: Remove `__MCF_TRY` and `__MCF_CATCH`
There is an issue about the traditional try-catch-rethrow approach: If the
rethrow operation leads to a call to `std::terminate()`, the stack will be
partially unwound, which makes debugging much harder.

This is solved by running cleanup functions in destructors, so exceptions
that will lead to `std::terminate()` are not caught. This also eliminates
the need to check for `-fno-exceptions`.

Signed-off-by: LIU Hao <lh_mouse@126.com>