xglobals,pch: Revert part of previous change
The crash that was referenced in the previous commit appears to a bug in
Clang. Let's not keep the inconsistency. This only has an impact on ARM64EC
code that calls these functions from an x64 DLL; all the other cases are
unaffected.

Ideally these functions could have been forwarded from NTDLL.DLL, but we
can't use a DEF file because of another bug in Clang in the x86-32 case.

Reference: https://github.com/llvm/llvm-project/issues/101355